Lexico:  -Escritura en fichero los errores.
Añadidos los cambios de Guille en el Sintactico.

Modificado el gitignore para que no quiera subir los archivos de la carpeta generated

+++ src/GestorTS.py
@@ -12,8 +12,11 @@ def creaTS():
 
 # Metodo que destruye la TS Actual y borra su identificador de la lista
 def destruyeTSActual():
-    symT.destroyTable(len(listaTS)-1)
-    del listaTS[len(listaTS)-1]
+    if symT.existTable(len(listaTS)-1):
+         symT.destroyTable(len(listaTS)-1)
+         del listaTS[len(listaTS)-1]
+    else:
+        print("No se puede borrar la tabla actual porque no existe")
 
 # Metodo que obtiene la TS Actual, sera la ultima insertada en la lista (por el final)
 def getTSActual():
@@ -21,9 +24,17 @@ def getTSActual():
 
 # Metodo que inserta el lexema en la TSActual
 # Retorna el ID asignado al lexema en esa tabla o falso si ya existia
-def insertaEnTSActual(lexema):
+def insertaLexemaEnTSActual(lexema):
     return symT.add(listaTS[len(listaTS)-1], lexema)
 
+# Metodo que inserta el tipo del lexema en la TSActual
+# Retorna falso si el lexema no existe en esa tabla o true si lo ha hecho correctamente
+def insertaTipoEnTSActual(lexema,tipo):
+    return symT.addType(listaTS[len(listaTS)-1], lexema, tipo)
+
+# Metodo que obtiene el tipo del lexema en la TSActual
+def getTipoLexemaTSActual(lexema):
+    return symT.checkType(len(listaTS)-1,lexema)
 
 # Metodo que busca en la TS actual un lexema y devuelve
 def buscaEnTSActual():
